+++ service/src/main/java/com/ericsson/eiffel/remrem/generate/config/SwaggerConfig.java
@@ -15,115 +15,40 @@
 package com.ericsson.eiffel.remrem.generate.config;
 
 import com.ericsson.eiffel.remrem.generate.constants.RemremGenerateServiceConstants;
-import io.swagger.annotations.Api;
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-
-import org.springframework.boot.actuate.autoconfigure.endpoint.web.CorsEndpointProperties;
-import org.springframework.boot.actuate.autoconfigure.endpoint.web.WebEndpointProperties;
-import org.springframework.boot.actuate.autoconfigure.web.server.ManagementPortType;
-import org.springframework.boot.actuate.endpoint.ExposableEndpoint;
-import org.springframework.boot.actuate.endpoint.web.EndpointLinksResolver;
-import org.springframework.boot.actuate.endpoint.web.EndpointMapping;
-import org.springframework.boot.actuate.endpoint.web.EndpointMediaTypes;
-import org.springframework.boot.actuate.endpoint.web.ExposableWebEndpoint;
-import org.springframework.boot.actuate.endpoint.web.WebEndpointsSupplier;
-import org.springframework.boot.actuate.endpoint.web.annotation.ControllerEndpointsSupplier;
-import org.springframework.boot.actuate.endpoint.web.annotation.ServletEndpointsSupplier;
-import org.springframework.boot.actuate.endpoint.web.servlet.WebMvcEndpointHandlerMapping;
+import org.springdoc.core.GroupedOpenApi;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.core.env.Environment;
-import org.springframework.util.StringUtils;
-
-import springfox.documentation.builders.ApiInfoBuilder;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.service.ApiInfo;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger.web.UiConfiguration;
-import springfox.documentation.swagger.web.UiConfiguration.Constants;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-import springfox.documentation.swagger.web.DocExpansion;
-import springfox.documentation.swagger.web.ModelRendering;
-import springfox.documentation.swagger.web.OperationsSorter;
-import springfox.documentation.swagger.web.TagsSorter;
-import springfox.documentation.swagger.web.UiConfiguration;
-import springfox.documentation.swagger.web.UiConfigurationBuilder;
 
 @Configuration
 public class SwaggerConfig {
 
+    @Value("${app.version:Unknown}")
+    private String appVersion;
+
     @Bean
-    public Docket api() {
-        return new Docket(DocumentationType.SWAGGER_2)
-                .select()
-                .apis(RequestHandlerSelectors.withClassAnnotation(Api.class))
-                .paths(PathSelectors.any())
-                .build()
-                .apiInfo(metaData());
-    }
-
-        @Bean
-        public UiConfiguration uiConfig() {
-
-            return UiConfigurationBuilder.builder()
-                .deepLinking(true)
-                .displayOperationId(false)
-                .defaultModelsExpandDepth(-1)
-                .defaultModelExpandDepth(1)
-                .defaultModelRendering(ModelRendering.EXAMPLE)
-                .displayRequestDuration(false)
-                .docExpansion(DocExpansion.NONE)
-                .filter(false)
-                .maxDisplayedTags(null)
-                .operationsSorter(OperationsSorter.ALPHA)
-                .showExtensions(false)
-                .tagsSorter(TagsSorter.ALPHA)
-                .supportedSubmitMethods(UiConfiguration.Constants.DEFAULT_SUBMIT_METHODS)
-                .validatorUrl(null)
-                .build();
-        }
-
-
-	@Bean
-        public WebMvcEndpointHandlerMapping webEndpointServletHandlerMapping(WebEndpointsSupplier webEndpointsSupplier,
-                ServletEndpointsSupplier servletEndpointsSupplier, ControllerEndpointsSupplier controllerEndpointsSupplier,
-                EndpointMediaTypes endpointMediaTypes, CorsEndpointProperties corsProperties,
-                WebEndpointProperties webEndpointProperties, Environment environment) {
-            List<ExposableEndpoint<?>> allEndpoints = new ArrayList();
-            Collection<ExposableWebEndpoint> webEndpoints = webEndpointsSupplier.getEndpoints();
-            allEndpoints.addAll(webEndpoints);
-            allEndpoints.addAll(servletEndpointsSupplier.getEndpoints());
-            allEndpoints.addAll(controllerEndpointsSupplier.getEndpoints());
-            String basePath = webEndpointProperties.getBasePath();
-            EndpointMapping endpointMapping = new EndpointMapping(basePath);
-            boolean shouldRegisterLinksMapping = this.shouldRegisterLinksMapping(webEndpointProperties, environment,
-                    basePath);
-            return new WebMvcEndpointHandlerMapping(endpointMapping, webEndpoints, endpointMediaTypes,
-                    corsProperties.toCorsConfiguration(), new EndpointLinksResolver(allEndpoints, basePath),
-                    shouldRegisterLinksMapping, null);
-        }
-
-        private boolean shouldRegisterLinksMapping(WebEndpointProperties webEndpointProperties, Environment environment,
-                String basePath) {
-            return webEndpointProperties.getDiscovery().isEnabled() && (StringUtils.hasText(basePath)
-                || ManagementPortType.get(environment).equals(ManagementPortType.DIFFERENT));
-        }
-
-    private ApiInfo metaData() {
+    public OpenAPI customOpenAPI() {
         final StringBuilder remremDescription = new StringBuilder();
         remremDescription.append("REMReM (REST Mailbox for Registered Messages) Generate "
                 + "for generating validated Eiffel messages.\n");
         remremDescription.append("<a href= " + RemremGenerateServiceConstants.DOCUMENTATION_URL + ">REMReM Generate documentation</a>");
 
-        return new ApiInfoBuilder()
-                .title("Eiffel REMReM Generate Service")
-                .description(remremDescription.toString())
+        return new OpenAPI()
+                .info(new Info()
+                        .title("Eiffel REMReM Generate Service")
+                        .description(remremDescription.toString())
+                        .version(appVersion))
+                .openapi("3.1.0");
+    }
+
+    @Bean
+    public GroupedOpenApi api() {
+        return GroupedOpenApi.builder()
+                .group("remrem-generate")
+                .pathsToMatch("/**")
                 .build();
     }
 }